+++ packages/supermassive-ast/src/TypedAST.ts
@@ -6,8 +6,6 @@
  * Our changes MUST be annotated inline for ease of future merging.
  */
 
-import { Maybe } from "../jsutils/Maybe";
-
 /**
  * [SUPERMASSIVE]
  *
@@ -188,7 +186,7 @@ export interface FieldNode {
 
 export interface ArgumentNode {
   readonly __type: TypeNode; // [SUPERMASSIVE] Add the value type
-  readonly __defaultValue?: Maybe<ValueNode>; // [SUPERMASSIVE] Add default value
+  readonly __defaultValue?: ValueNode | null; // [SUPERMASSIVE] Add default value
   readonly kind: "Argument";
   readonly loc?: Location;
   readonly name: NameNode;